+++ packages/rx_bloc_cli/lib/src/package_version.dart
@@ -1,1 +1,1 @@
-const packageVersion = '0.0.4';
+const packageVersion = '0.0.5';