+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/bin/server/config.dart
@@ -1,0 +1,2 @@
+/// TODO: In order to use FCM and send push messages to target devices, you need to add your server key
+const firebasePushServerKey = '';