+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/lib/base/app/{{project_name}}.dart
@@ -5,18 +5,24 @@
 // license that can be found in the LICENSE file or at
 // https://opensource.org/licenses/MIT.
 
-{{#analytics}}
-import 'package:firebase_analytics/observer.dart';
-{{/analytics}}
+import 'package:dio/dio.dart';{{#analytics}}
+import 'package:firebase_analytics/observer.dart';{{/analytics}}{{#push_notifications}}
+import 'package:firebase_messaging/firebase_messaging.dart';{{/push_notifications}}
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:provider/provider.dart';
 
-import '../../l10n/l10n.dart';
+import '../../l10n/l10n.dart';{{#analytics}}
+import '../data_sources/remote/interceptors/analytics_interceptor.dart';{{/analytics}}
+import '../data_sources/remote/interceptors/auth_interceptor.dart';
 import '../di/app_dependencies.dart';
 import '../routers/router.gr.dart' as router;
 import '../theme/design_system.dart';
-import 'config/environment_config.dart';
+import '../utils/helpers.dart';
+import 'config/app_constants.dart';
+import 'config/environment_config.dart';{{#push_notifications}}
+import 'initialization/firebase_messaging_callbacks.dart';{{/push_notifications}}
 
 /// This widget is the root of your application.
 class {{#pascalCase}}{{project_name}}{{/pascalCase}} extends StatelessWidget {
@@ -27,15 +33,58 @@ class {{#pascalCase}}{{project_name}}{{/pascalCase}} extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) => MultiProvider(
-        providers: AppDependencies.of(context).providers,
+        providers: AppDependencies.of(context,config).providers,
         child: _MyMaterialApp(_router),
       );
 }
 
-class _MyMaterialApp extends StatelessWidget {
+/// Wrapper around the MaterialApp widget to provide additional functionality
+/// accessible throughout the app (such as App-level dependencies, Firebase
+/// services, etc).
+class _MyMaterialApp extends StatefulWidget {
   const _MyMaterialApp(this._router);
 
   final router.Router _router;
+
+@override
+__MyMaterialAppState createState() => __MyMaterialAppState();
+}
+
+class __MyMaterialAppState extends State<_MyMaterialApp> {
+
+  @override
+  void initState() { {{#push_notifications}}
+    _configureFCM(); {{/push_notifications}}
+    _addInterceptors();
+
+    super.initState();
+  }{{#push_notifications}}
+
+  Future<void> _configureFCM() async {
+    /// Initialize the FCM callbacks
+    if (kIsWeb){
+        await safeRun(
+            () => FirebaseMessaging.instance.getToken(vapidKey: webVapidKey));
+    }
+    FirebaseMessaging.instance
+        .getInitialMessage()
+        .then((message) => onInitialMessageOpened(context, message));
+    FirebaseMessaging.instance.onTokenRefresh
+        .listen((token) => onFCMTokenRefresh(context, token));
+    FirebaseMessaging.onMessage
+        .listen((message) => onForegroundMessage(context, message));
+    FirebaseMessaging.onMessageOpenedApp
+        .listen((message) => onMessageOpenedFromBackground(context, message));
+  }{{/push_notifications}}
+
+  void _addInterceptors(){
+    context.read<Dio>().interceptors.addAll([
+      AuthInterceptor(context.read(), context.read(), context.read()),{{#analytics}}
+      AnalyticsInterceptor(context.read()),{{/analytics}}
+
+      /// TODO: Add your own interceptors here
+    ]);
+  }
 
   @override
   Widget build(BuildContext context) => MaterialApp.router(
@@ -47,11 +96,12 @@ class _MyMaterialApp extends StatelessWidget {
           GlobalMaterialLocalizations.delegate,
         ],
         supportedLocales: AppLocalizations.supportedLocales,
-        routeInformationParser: _router.defaultRouteParser(),
-        routerDelegate: _router.delegate({{#analytics}}
+        routeInformationParser: widget._router.defaultRouteParser(),
+        routerDelegate: widget._router.delegate({{#analytics}}
           navigatorObservers: () => [
-            Provider.of<FirebaseAnalyticsObserver>(context, listen: false),
+            context.read<FirebaseAnalyticsObserver>(),
           ],
         {{/analytics}}),
+        debugShowCheckedModeBanner: false,
       );
 }