+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/lib/base/common_use_cases/fetch_access_token_use_case.dart
@@ -3,16 +3,16 @@ import '../repositories/auth_repository.dart';
 /// Returns access token, if is saved, by default.
 class FetchAccessTokenUseCase {
   FetchAccessTokenUseCase(
-      this._authRepository,
-      );
+    this._authRepository,
+  );
 
   final AuthRepository _authRepository;
 
   /// If you want to get refresh token, you should set forceFetchNewToken to true
   Future<String?> execute({bool forceFetchNewToken = false}) async {
-    if(!forceFetchNewToken){
+    if (!forceFetchNewToken) {
       final token = await _authRepository.getToken();
-      if (token != null){
+      if (token != null) {
         return token;
       }
     }