+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/lib/base/repositories/auth_repository.dart
@@ -6,28 +6,41 @@
 // https://opensource.org/licenses/MIT.
 
 import '../data_sources/local/auth_token_data_source.dart';
+import '../data_sources/remote/auth_data_source.dart';
+import '../models/auth_token_model.dart';
+import '../models/request_models/authenticate_user_request_model.dart';
 
 class AuthRepository {
-  AuthRepository(this._authDataSource);
+  AuthRepository(
+      {required AuthTokenDataSource authTokenDataSource,
+      required AuthDataSource authDataSource})
+      : _authDataSource = authDataSource,
+        _authTokenDataSource = authTokenDataSource;
 
-  final AuthTokenDataSource _authDataSource;
+  final AuthTokenDataSource _authTokenDataSource;
+  final AuthDataSource _authDataSource;
 
   // Get token string if there is saved
-  Future<String?> getToken() => _authDataSource.getToken();
+  Future<String?> getToken() => _authTokenDataSource.getToken();
+
+  Future<bool> isAuthenticated() async {
+    final token = await getToken();
+    return token != null;
+  }
 
   // Persist new token string in secure storage
   Future<void> saveToken(String newToken) =>
-      _authDataSource.saveToken(newToken);
+      _authTokenDataSource.saveToken(newToken);
 
   // Get refreshToken string if there is saved
-  Future<String?> getRefreshToken() => _authDataSource.getRefreshToken();
+  Future<String?> getRefreshToken() => _authTokenDataSource.getRefreshToken();
 
   // Persist new refreshToken string in secure storage
   Future<void> saveRefreshToken(String newRefreshToken) =>
-      _authDataSource.saveRefreshToken(newRefreshToken);
+      _authTokenDataSource.saveRefreshToken(newRefreshToken);
 
   // Delete all saved tokens
-  Future<void> clearAuthData() => _authDataSource.clear();
+  Future<void> clearAuthData() => _authTokenDataSource.clear();
 
 // Fetch new access token
   Future<String?> fetchNewToken() async {
@@ -40,4 +53,11 @@ class AuthRepository {
       return null;
     }
   }
+
+  Future<AuthTokenModel> authenticate(
+          {String? email, String? password, String? refreshToken}) =>
+      _authDataSource.authenticate(AuthUserRequestModel(
+          username: email, password: password, refreshToken: refreshToken));
+
+  Future<void> logout() => _authDataSource.logout();
 }